\name{fit}
\alias{fit}
\title{Fit Particle Interaction Models}
\description{
Generic function for fitting particle interaction models to data.
}
\usage{
fit(object, ...)
}
\arguments{
\item{object}{Object containing data to be fit}
\item{...}{Additional arguments passed to fitting methods}
}
\seealso{
\code{\link{fit.particle.data}} for fitting particle data models
}
\keyword{methods}